move modules to services directory
